fix: Long to Integer

+++ src/main/java/rocketseat/com/passin/repositories/CheckinRepository.java
@@ -3,4 +3,4 @@ package rocketseat.com.passin.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import rocketseat.com.passin.domain.checkin.Checkin;
 
-public interface CheckinRepository extends JpaRepository<Checkin, Long> {}
+public interface CheckinRepository extends JpaRepository<Checkin, Integer> {}
